Basic validation and refactoring
Created very basic input validation for the first window that is
presented for the user to input their name, IP, and Port. If there is no
content for one or more of the fields, the screen will not allow one to
continue until all are complete.

Also changed the name of some functions to fit the language spec.