Port verification and minor fixes
Added verification for the port that the user entered into the form.
checked to make sure that it was an unsigned 16 bit value.

Also modified some of the code in terms of accessing the port that the
user put in. Originally there was a constant `PORT` that took the value
from the text field and casted it to an integer. This would cause the
program to panic when the user input a letter. I delegated this task of
casting to an integer to the validation function which would then
determine whether an error should be displayed.

Additionally cleaned up the IP Address verification code so that it was
less verbose.